+++ src/Modules/PassportGenerator/PassportGeneratorControl.php
@@ -2,46 +2,31 @@
 
 namespace Foodsharing\Modules\PassportGenerator;
 
-use Endroid\QrCode\QrCode;
-use Foodsharing\Modules\Bell\BellGateway;
-use Foodsharing\Modules\Bell\DTO\Bell;
 use Foodsharing\Modules\Core\Control;
-use Foodsharing\Modules\Core\DBConstants\Bell\BellType;
-use Foodsharing\Modules\Core\DBConstants\Foodsaver\Gender;
-use Foodsharing\Modules\Core\DBConstants\Foodsaver\Role;
-use Foodsharing\Modules\Foodsaver\FoodsaverGateway;
 use Foodsharing\Modules\Region\RegionGateway;
-use Foodsharing\Modules\Uploads\UploadsTransactions;
 use Foodsharing\Utility\IdentificationHelper;
-use setasign\Fpdi\Tcpdf\Fpdi;
 
 final class PassportGeneratorControl extends Control
 {
 	private $regionId;
 	private $region;
-	private BellGateway $bellGateway;
 	private RegionGateway $regionGateway;
 	private PassportGeneratorGateway $passportGeneratorGateway;
-	private FoodsaverGateway $foodsaverGateway;
 	private IdentificationHelper $identificationHelper;
-	private UploadsTransactions $uploadsTransactions;
+	private PassportGeneratorTransaction $passportGeneratorTransaction;
 
 	public function __construct(
 		PassportGeneratorView $view,
-		BellGateway $bellGateway,
 		RegionGateway $regionGateway,
 		PassportGeneratorGateway $passportGateway,
-		FoodsaverGateway $foodsaverGateway,
 		IdentificationHelper $identificationHelper,
-		UploadsTransactions $uploadsTransactions
+		PassportGeneratorTransaction $passportGeneratorTransaction
 	) {
 		$this->view = $view;
-		$this->bellGateway = $bellGateway;
 		$this->regionGateway = $regionGateway;
 		$this->passportGeneratorGateway = $passportGateway;
-		$this->foodsaverGateway = $foodsaverGateway;
 		$this->identificationHelper = $identificationHelper;
-		$this->uploadsTransactions = $uploadsTransactions;
+		$this->passportGeneratorTransaction = $passportGeneratorTransaction;
 
 		parent::__construct();
 
@@ -69,7 +54,7 @@ final class PassportGeneratorControl extends Control
 		$this->pageHelper->addTitle($this->translator->trans('pass.bread'));
 
 		if (isset($_POST['passes']) && !empty($_POST['passes'])) {
-			$this->generate($_POST['passes']);
+			$this->passportGeneratorTransaction->generate($_POST['passes'], true, $this->region);
 		}
 
 		if ($regions = $this->passportGeneratorGateway->getPassFoodsaver($this->regionId)) {
@@ -105,204 +90,6 @@ final class PassportGeneratorControl extends Control
 		if (isset($_GET['dl2'])) {
 			$this->download2();
 		}
-	}
-
-	public function generate(array $foodsavers): void
-	{
-		$tmp = [];
-		foreach ($foodsavers as $foodsaver) {
-			$tmp[$foodsaver] = (int)$foodsaver;
-		}
-		$foodsavers = $tmp;
-		$is_generated = [];
-
-		$pdf = new Fpdi();
-		$pdf->AddPage();
-		$pdf->SetTextColor(0, 0, 0);
-		$pdf->AddFont('Ubuntu-L', '', 'lib/font/ubuntul.php', true);
-		$pdf->AddFont('AcmeFont Regular', '', 'lib/font/acmefont.php', true);
-
-		$x = 0;
-		$y = 0;
-		$card = 0;
-
-		$noPhoto = [];
-
-		end($foodsavers);
-
-		$pdf->setSourceFile('img/foodsharing_logo.pdf');
-		$fs_logo = $pdf->importPage(1);
-
-		foreach ($foodsavers as $fs_id) {
-			if ($foodsaver = $this->foodsaverGateway->getFoodsaverDetails($fs_id)) {
-				if (empty($foodsaver['photo'])) {
-					$noPhoto[] = $foodsaver['name'] . ' ' . $foodsaver['nachname'];
-
-					$bellData = Bell::create(
-						'passgen_failed_title',
-						'passgen_failed',
-						'fas fa-camera',
-						['href' => '/?page=settings'],
-						['user' => $this->session->user('name')],
-						BellType::createIdentifier(BellType::PASS_CREATION_FAILED, $foodsaver['id'])
-					);
-					$this->bellGateway->addBell($foodsaver['id'], $bellData);
-					//continue;
-				}
-
-				$pdf->SetTextColor(0, 0, 0);
-
-				++$card;
-
-				$this->passportGeneratorGateway->passGen($this->session->id(), $foodsaver['id']);
-
-				$pdf->Image('img/pass_bg.png', 10 + $x, 10 + $y, 83, 55);
-
-				$pdf->SetFont('Ubuntu-L', '', 10);
-				$name = $foodsaver['name'] . ' ' . $foodsaver['nachname'];
-				$maxWidth = 49;
-				if ($pdf->GetStringWidth($name) > $maxWidth) {
-					$pdf->SetFont('Ubuntu-L', '', 8);
-					if ($pdf->GetStringWidth($name) <= $maxWidth) {
-						$pdf->Text(41 + $x, 30 + $y, $name);
-					}
-					$size = 8;
-					while ($pdf->GetStringWidth($foodsaver['name']) > $maxWidth
-						|| $pdf->GetStringWidth($foodsaver['nachname']) > $maxWidth
-					) {
-						$size -= 0.5;
-						$pdf->SetFont('Ubuntu-L', '', $size);
-					}
-					$pdf->Text(41 + $x, 30.2 + $y, $foodsaver['name']);
-					$pdf->Text(41 + $x, 33.2 + $y, $foodsaver['nachname']);
-				} else {
-					$pdf->Text(41 + $x, 30 + $y, $name);
-				}
-				$pdf->SetFont('Ubuntu-L', '', 10);
-				$pdf->Text(41 + $x, 39 + $y, $this->getRole($foodsaver['geschlecht'], $foodsaver['rolle']));
-				$pdf->Text(41 + $x, 48 + $y, date('d. m. Y', time() - 1814400));
-				$pdf->Text(41 + $x, 57 + $y, date('d. m. Y', time() + 94608000));
-
-				$pdf->SetFont('Ubuntu-L', '', 6);
-				$pdf->Text(41 + $x, 28 + $y, 'Name');
-				$pdf->Text(41 + $x, 37 + $y, 'Rolle');
-				$pdf->Text(41 + $x, 46 + $y, 'Gültig ab');
-				$pdf->Text(41 + $x, 55 + $y, 'Gültig bis');
-
-				$pdf->SetFont('Ubuntu-L', '', 9);
-				$pdf->SetTextColor(255, 255, 255);
-				$pdf->SetXY(40 + $x, 13.2 + $y);
-				$pdf->Cell(50, 5, 'ID ' . $fs_id, 0, 0, 'R');
-
-				$pdf->SetFont('AcmeFont Regular', '', 5.3);
-				$pdf->Text(12.8 + $x, 18.6 + $y, $this->translator->trans('pass.claim'));
-
-				$pdf->useTemplate($fs_logo, 13.5 + $x, 13.6 + $y, 29.8);
-
-				$style = [
-					'vpadding' => 'auto',
-					'hpadding' => 'auto',
-					'fgcolor' => [0, 0, 0],
-					'bgcolor' => false, // array(255,255,255)
-					'module_width' => 1, // width of a single module in points
-					'module_height' => 1 // height of a single module in points
-				];
-
-				// FIXME Do we really always want fs.de here?!
-				// QRCODE,L : QR-CODE Low error correction
-				$pdf->write2DBarcode('https://foodsharing.de/profile/' . $fs_id, 'QRCODE,L', 70.5 + $x, 43 + $y, 20, 20, $style, 'N');
-
-				if ($photo = $this->foodsaverGateway->getPhotoFileName($fs_id)) {
-					if (str_starts_with($photo, '/api/uploads')) {
-						// get the UUID and create a resized file
-						$uuid = substr($photo, strlen('/api/uploads/'));
-						$filename = $this->uploadsTransactions->getFileLocation($uuid, 200, 257, 0);
-						if (!file_exists($filename)) {
-							$originalFilename = $this->uploadsTransactions->getFileLocation($uuid);
-							$this->uploadsTransactions->resizeImage($originalFilename, $filename, 200, 257, 0);
-						}
-						$pdf->Image($filename, 14 + $x, 29.7 + $y, 24);
-					} else {
-						if (file_exists('images/crop_' . $photo)) {
-							$pdf->Image('images/crop_' . $photo, 14 + $x, 29.7 + $y, 24);
-						} elseif (file_exists('images/' . $photo)) {
-							$pdf->Image('images/' . $photo, 14 + $x, 29.7 + $y, 22);
-						}
-					}
-				}
-
-				if ($x == 0) {
-					$x += 95;
-				} else {
-					$y += 65;
-					$x = 0;
-				}
-
-				if ($card == 8) {
-					$card = 0;
-					$pdf->AddPage();
-					$x = 0;
-					$y = 0;
-				}
-
-				$is_generated[] = $foodsaver['id'];
-			}
-		}
-		if (!empty($noPhoto)) {
-			$this->flashMessageHelper->info(
-				$this->translator->trans('pass.noPhoto')
-				. join(', ', $noPhoto)
-				. $this->translator->trans('pass.notGenerated')
-			);
-		}
-
-		$this->passportGeneratorGateway->updateLastGen($is_generated);
-
-		$bez = strtolower($this->region['name']);
-
-		$bez = str_replace(['ä', 'ö', 'ü', 'ß'], ['ae', 'oe', 'ue', 'ss'], $bez);
-		$bez = preg_replace('/[^a-zA-Z]/', '', $bez);
-
-		$pdf->Output('foodsaver_pass_' . $bez . '.pdf', 'D');
-		exit();
-	}
-
-	public function getRole(int $gender_id, int $role_id): string
-	{
-		switch ($gender_id) {
-			case Gender::MALE:
-			  $roles = [
-					Role::FOODSHARER => $this->translator->trans('terminology.foodsharer.m'),
-					Role::FOODSAVER => $this->translator->trans('terminology.foodsaver.m'),
-					Role::STORE_MANAGER => $this->translator->trans('terminology.storemanager.m'),
-					Role::AMBASSADOR => $this->translator->trans('terminology.ambassador.m'),
-					Role::ORGA => $this->translator->trans('terminology.ambassador.m'),
-				];
-				break;
-
-			case Gender::FEMALE:
-			  $roles = [
-					Role::FOODSHARER => $this->translator->trans('terminology.foodsharer.f'),
-					Role::FOODSAVER => $this->translator->trans('terminology.foodsaver.f'),
-					Role::STORE_MANAGER => $this->translator->trans('terminology.storemanager.f'),
-					Role::AMBASSADOR => $this->translator->trans('terminology.ambassador.f'),
-					Role::ORGA => $this->translator->trans('terminology.ambassador.f'),
-				];
-				break;
-
-			// All others
-			default:
-				$roles = [
-					Role::FOODSHARER => $this->translator->trans('terminology.foodsharer.d'),
-					Role::FOODSAVER => $this->translator->trans('terminology.foodsaver.d'),
-					Role::STORE_MANAGER => $this->translator->trans('terminology.storemanager.d'),
-					Role::AMBASSADOR => $this->translator->trans('terminology.ambassador.d'),
-					Role::ORGA => $this->translator->trans('terminology.ambassador.d'),
-				];
-			  break;
-		}
-
-		return $roles[$role_id];
 	}
 
 	private function download1(): void